Make dropdown not deprecated, add additional prop 'initiallyOpened'
