+++ src/components/dropdowns/_dropdowns.scss
@@ -5,7 +5,6 @@
   align-items: center;
   min-height: spacing(l);
   height: spacing(l);
-  width: 100%;
   padding: 0 spacing(s);
   font-size: fontSize(medium);
   line-height: rhythm(5 / 6);
