+++ src/index.js
@@ -26,9 +26,7 @@ export {
 export {
   default as ContentBoxTitle,
 } from './components/content-box/ContentBoxTitle';
-export {
-  default as DropdownDeprecated,
-} from './components/dropdowns-deprecated/DropdownDeprecated';
+export {default as Dropdown} from './components/dropdowns/Dropdown';
 export {
   default as FlashMessage,
 } from './components/flash-messages/FlashMessage';
