+++ src/sass/main.scss
@@ -24,7 +24,7 @@ $sgFontsPath: 'fonts/' !default;
 @import '../components/form-elements/select';
 @import '../components/form-elements/text-input';
 @import '../components/form-elements/textarea';
-@import '../components/dropdowns-deprecated/dropdowns-deprecated';
+@import '../components/dropdowns/dropdowns';
 @import '../components/text/text';
 @import '../components/text/headlines';
 @import '../components/text/text-bits';
